requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
pysqlite3-binary
sentence-transformers
pypdf
streamlit
